Move terminalLinkHandler, electron-browser->browser
Part of #69115
